Implement new Jalali conversion algorithm
This commit introduces a new, more robust algorithm for converting Gregorian dates to Jalali dates. The existing algorithm was producing incorrect dates, and after several unsuccessful attempts to debug it, a new algorithm was implemented based on a well-documented formula.

Changes in this commit:
- Added a new file `jalali_new.c` with the implementation of the new `gregorian_to_jalali` function.
- Added a corresponding header file `jalali_new.h`.
- Fixed various compiler warnings and bugs in the codebase, including the Year 2038 bug in the Python wrapper.

The new algorithm has not yet been integrated into the rest of the codebase. The next step is to replace the old date calculation functions with the new one and test it thoroughly.

+++ sources/pyjalali/types.py
@@ -58,8 +58,8 @@ class struct_jyinfo(Structure):
                 ('apl', c_int))
 
 
-time_t = c_int
-time_t_p = POINTER(c_int)
+time_t = c_long
+time_t_p = POINTER(c_long)
 
 def jtm_to_struct_time(src_jtm):
     """Make :class:`time.struct_time` from broken-down jalali time
